pyproject.toml
[project]
name = "rag-qa-engine"
version = "0.1.0"
description = "RAG QA Engine - Portfolio Project"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faiss-cpu>=1.13.2",
    "fastapi>=0.135.0",
    "groq>=1.0.0",
    "jinja2>=3.1.6",
    "nltk>=3.9.3",
    "pdfplumber>=0.11.9",
    "pymupdf>=1.27.1",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "pyyaml>=6.0.3",
    "rank-bm25>=0.2.2",
    "sentence-transformers>=5.2.3",
    "torch>=2.10.0",
    "uvicorn>=0.41.0",
]

[tool.uv]
index-url = "https://pypi.org/simple"

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[tool.uv.sources]
torch = [{ index = "pytorch-cpu" }]
